Unification (#28)
* Unify templated files

A lot of files are identical throughout the different packages.
This is the base for unifying them for easier handling.

* Adjust tests path

Some packages use 'tests' as tests directory, while other use 'Tests'. The reference in phpunit.xml.dist is adjusted accordingly.

* Build: Ignore PHP 5.4 and 5.5

The images used for tests with these versions have problems with using composer (giving curl error 60), so they are removed for now, until a solution is found.